Pushed Initial File to Github
This file also contains the compression algorithm, just so that the file can be compressed and decompressed. This is because a file that is compressed and decompressed should result in the exact same number of bytes as the original file - therefore, both compression and decompression are included to check the verity of my methods.